feat(dtt): :art: decoupling dtt error
